:truck: Move jobs to jobs application
