+++ creator/jobs/schema.py
@@ -6,7 +6,35 @@ from graphene_django.filter import DjangoFilterConnectionField
 from django_filters import FilterSet, OrderingFilter
 from graphql import GraphQLError
 
-from creator.jobs.models import JobLog
+from creator.jobs.models import Job, JobLog
+
+
+class JobFilter(FilterSet):
+    order_by = OrderingFilter(fields=("created_on", "last_run", "name"))
+
+    class Meta:
+        model = Job
+        fields = ["name", "active", "failing"]
+
+
+class JobNode(DjangoObjectType):
+    enqueued_at = graphene.DateTime()
+
+    class Meta:
+        model = Job
+        interfaces = (graphene.relay.Node,)
+
+    @classmethod
+    def get_node(cls, info, name):
+        """
+        Only return node if user is admin
+        """
+        user = info.context.user
+
+        if not user.has_perm("jobs.view_job"):
+            return Job.objects.none()
+
+        return Job.objects.get(name=name)
 
 
 class JobLogFilter(FilterSet):
@@ -47,6 +75,21 @@ class JobLogNode(DjangoObjectType):
 
 
 class Query(object):
+    job = relay.Node.Field(JobNode, description="Get a single job")
+    all_jobs = DjangoFilterConnectionField(
+        JobNode, filterset_class=JobFilter, description="Get job statuses"
+    )
+
+    def resolve_all_jobs(self, info, **kwargs):
+        """
+        Return all Jobs
+        """
+        user = info.context.user
+        if not user.has_perm("jobs.list_all_job"):
+            raise GraphQLError("Not allowed")
+
+        return Job.objects.all()
+
     job_log = relay.Node.Field(JobLogNode, description="Get a single job log")
     all_job_logs = DjangoFilterConnectionField(
         JobLogNode,
@@ -56,6 +99,7 @@ class Query(object):
 
     def resolve_all_job_logs(self, info, **kwargs):
         """
+        Return all Job Logs
         """
         user = info.context.user
         if not user.has_perm("jobs.list_all_joblog"):
